# app/demo/docs/swagger.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo",
    "version": "v0.1.0"
  },
...
  "host": "localhost:5188",
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/group/add": {
      "post": {
        "summary": "添加群成员",
        "operationId": "Group_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apigroupAddRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apigroupAddReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/addManager": {
      "post": {
        "summary": "添加管理员",
        "operationId": "Group_AddManager",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupAddManagerRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupAddManagerReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/create": {
      "post": {
        "summary": "创建群组",
        "operationId": "Group_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupCreateRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupCreateReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/dismiss": {
      "post": {
        "summary": "解散群组",
        "operationId": "Group_Dismiss",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupDismissRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupDismissReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/info": {
      "post": {
        "summary": "群信息",
        "operationId": "Group_Info",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupInfoRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupInfoReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/join": {
      "post": {
        "summary": "加入群组",
        "operationId": "Group_Join",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupJoinRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupJoinReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/kick": {
      "post": {
        "summary": "退出群组",
        "operationId": "Group_Kick",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupKickRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupKickReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/memberList": {
      "post": {
        "summary": "获取群成员列表",
        "operationId": "Group_MemberList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupMemberListRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupMemberListReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/quit": {
      "post": {
        "summary": "退出群组",
        "operationId": "Group_Quit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupQuitRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupQuitReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/removeManager": {
      "post": {
        "summary": "删除管理员",
        "operationId": "Group_RemoveManager",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupRemoveManagerRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupRemoveManagerReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/rename": {
      "post": {
        "summary": "重命名群组",
        "operationId": "Group_Rename",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupRenameRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupRenameReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/setAvatar": {
      "post": {
        "summary": "设置群组头像",
        "operationId": "Group_SetAvatar",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupSetAvatarRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupSetAvatarReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
    "/group/setDisplayName": {
      "post": {
        "summary": "设置群名片",
        "operationId": "Group_SetDisplayName",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/groupSetDisplayNameRsp"
            }
          },
          "401": {
            "description": "Returned when the user does not login.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "403": {
            "description": "Returned when the user does not have permission to access the resource.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "404": {
            "description": "Returned when the resource does not exist.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "599": {
            "description": "Returned when the business error happen.",
            "schema": {
              "$ref": "#/definitions/swaggerResponseBody"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/groupSetDisplayNameReq"
            }
          }
        ],
        "tags": [
          "群组"
        ]
      }
    },
...
  },
  "definitions": {
...
    "apigroupAddReq": {
      "type": "object",
      "properties": {
        "group_id": {
          "type": "string",
          "format": "int64"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        }
      }
    },
    "apigroupAddRsp": {
      "type": "object"
    },
...
    "groupAddManagerReq": {
      "type": "object",
      "properties": {
        "group_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "groupAddManagerRsp": {
      "type": "object"
    },
    "groupCreateReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "int64"
          }
        },
        "avatar": {
          "type": "string"
        },
        "notice": {
          "type": "string"
        },
        "intro": {
          "type": "string"
        }
      }
    },
    "groupCreateRsp": {
      "type": "object",
      "properties": {
        "group_id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
    "groupDismissRsp": {
      "type": "object"
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "swaggerResponseBody": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "detail": {
          "type": "string"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
}